backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
aiohttp==3.10.10
python-dotenv==1.0.1